% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{build_table_s3_gt}
\alias{build_table_s3_gt}
\title{Build a gt Sensitivity Analysis Table (S3 Table)}
\usage{
build_table_s3_gt(sensitivity_results)
}
\arguments{
\item{sensitivity_results}{Tibble from \code{\link[=build_sensitivity_results]{build_sensitivity_results()}}.}
}
\value{
A \code{gt_tbl} object.
}
\description{
Build a gt Sensitivity Analysis Table (S3 Table)
}
\examples{
\dontrun{
results <- build_sensitivity_results(files)
gt_s3 <- build_table_s3_gt(results)
}

}
