% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_flexneri_sf2a}
\alias{combined_flexneri_sf2a}
\alias{combined_flexneri_sf3a}
\title{Combined S. flexneri Model Fit: Sf2a OSP}
\format{
An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1600 rows and 7 columns.

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1600 rows and 7 columns.
}
\source{
icddr,b / UCD-SERG SOSAR cohort, Dhaka 2021-2022.
Mock version: 5 subjects, 20 iterations, 2 chains.
}
\usage{
combined_flexneri_sf2a

combined_flexneri_sf3a
}
\description{
Pools \emph{S. flexneri} 2a and 3a infected individuals, leveraging
O-antigen cross-reactivity.
}
\keyword{datasets}
