% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dL_clean_Ipab_new_under5}
\alias{dL_clean_Ipab_new_under5}
\alias{dL_clean_Ipab_new_plus5}
\title{Case Data: Age-Stratified IpaB}
\format{
An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 12 columns.

An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30 rows and 12 columns.
}
\source{
SOSAR cohort, Dhaka 2021-2022. Mock version: 5 subjects.
}
\usage{
dL_clean_Ipab_new_under5

dL_clean_Ipab_new_plus5
}
\description{
Case Data: Age-Stratified IpaB
}
\keyword{datasets}
