% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dL_combined_flexneri_sf2a_case}
\alias{dL_combined_flexneri_sf2a_case}
\alias{dL_combined_flexneri_sf3a_case}
\title{Case Data: Combined Flexneri}
\format{
An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 40 rows and 12 columns.

An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 40 rows and 12 columns.
}
\source{
SOSAR cohort, Dhaka 2021-2022. Mock version: 5 subjects.
}
\usage{
dL_combined_flexneri_sf2a_case

dL_combined_flexneri_sf3a_case
}
\description{
Case Data: Combined Flexneri
}
\keyword{datasets}
