% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dL_serotype_sf2a}
\alias{dL_serotype_sf2a}
\alias{dL_serotype_sf3a}
\alias{dL_serotype_sonnei}
\title{Case Data: Serotype-Restricted}
\format{
An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 30 rows and 12 columns.

An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 12 columns.

An object of class \code{case_data} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 12 columns.
}
\source{
SOSAR cohort, Dhaka 2021-2022. Mock version: 5 subjects.
}
\usage{
dL_serotype_sf2a

dL_serotype_sf3a

dL_serotype_sonnei
}
\description{
Case Data: Serotype-Restricted
}
\keyword{datasets}
