% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_individual_obs.R
\name{extract_individual_obs}
\alias{extract_individual_obs}
\title{Observed individual series (long) for grey background trajectories}
\usage{
extract_individual_obs(raw_dataset)
}
\arguments{
\item{raw_dataset}{Case/raw data.}
}
\value{
Tibble \verb{id, t, value, iso}.
}
\description{
Observed individual series (long) for grey background trajectories
}
