% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_s2_age_spaghetti.R
\name{figure_s2_age_spaghetti}
\alias{figure_s2_age_spaghetti}
\title{S2 Fig: raw longitudinal trajectories, children aged >= 5 years}
\usage{
figure_s2_age_spaghetti(case_list_plus5)
}
\arguments{
\item{case_list_plus5}{Named list of the five \verb{>=5y} single-antigen case
datasets.}
}
\value{
A ggplot.
}
\description{
S2 Fig: raw longitudinal trajectories, children aged >= 5 years
}
