% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_with_individuals.R
\name{figure_with_individuals}
\alias{figure_with_individuals}
\title{Population-trajectory panel for a single antigen (B-E of Fig 5)}
\usage{
figure_with_individuals(
  model_obj,
  raw_dataset,
  antigen_label,
  t_grid = seq(0, 210, by = 5),
  line_color = "#1f77b4",
  ribbon_alpha = 0.15,
  individual_alpha = 0.12,
  individual_color = "grey70",
  log_y = TRUE,
  xlim = c(0, 210)
)
}
\arguments{
\item{model_obj}{Fitted \code{sr_model}.}

\item{raw_dataset}{Case data for the grey individual lines.}

\item{antigen_label}{Panel title.}

\item{t_grid, line_color, ribbon_alpha, individual_alpha, individual_color, log_y, xlim}{Styling controls.}
}
\value{
A ggplot.
}
\description{
Population-trajectory panel for a single antigen (B-E of Fig 5)
}
