% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fit_obj.R
\name{load_fit_obj}
\alias{load_fit_obj}
\title{Load the \code{fit_obj} object from a single \code{sensitivity_*.rda}}
\usage{
load_fit_obj(path)
}
\arguments{
\item{path}{File path.}
}
\value{
The fitted \code{sr_model}.
}
\description{
Load the \code{fit_obj} object from a single \code{sensitivity_*.rda}
}
