% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_durdia}
\alias{mock_durdia}
\title{Mock Duration of Diarrhea Data}
\format{
A tibble with columns \code{CaseID} and \code{DurDia_hours}.
}
\usage{
mock_durdia
}
\description{
A tibble mimicking the diarrhea duration spreadsheet.
}
\keyword{datasets}
