% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock_metadata}
\alias{mock_metadata}
\title{Mock Participant Metadata}
\format{
A tibble with 10 columns: \code{sampleID}, \code{Gender}, \code{HhIncome},
\code{DiaBlood}, \code{DiaWatery}, \code{Fev}, \code{StoolNo}, \code{MUAC}, \code{IVFluidRec},
\code{HosDur}.
}
\usage{
mock_metadata
}
\description{
A tibble mimicking the additional metadata spreadsheet.
}
\keyword{datasets}
