% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overall_IpaB_pop_6}
\alias{overall_IpaB_pop_6}
\alias{overall_Sf2a_pop_6}
\alias{overall_Sf3a_pop_6}
\alias{overall_Sf6_pop_6}
\alias{overall_Sonnei_pop_6}
\title{Overall Hierarchical Model Fit: IpaB}
\format{
An \code{sr_model} tibble with columns \code{Iteration}, \code{Chain},
\code{Parameter}, \code{Iso_type}, \code{Stratification}, \code{Subject}, \code{value}.
Key attributes:
\describe{
\item{\code{population_params}}{Tibble of \code{mu.par} and \code{prec.logy}
population-level MCMC draws.}
\item{\code{nChains}, \code{nIterations}}{MCMC dimensions.}
}

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2000 rows and 7 columns.

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2000 rows and 7 columns.

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2000 rows and 7 columns.

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2000 rows and 7 columns.
}
\source{
icddr,b / UCD-SERG SOSAR cohort, Dhaka 2021-2022.
Mock version: 5 subjects, 20 iterations, 2 chains.
}
\usage{
overall_IpaB_pop_6

overall_Sf2a_pop_6

overall_Sf3a_pop_6

overall_Sf6_pop_6

overall_Sonnei_pop_6
}
\description{
Fitted Bayesian hierarchical two-phase antibody kinetic model for
invasion plasmid antigen B (IpaB), pooling all confirmed-\emph{Shigella}
cases regardless of infecting serotype.
}
\keyword{datasets}
