% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overall_IpaB_pop_plus5}
\alias{overall_IpaB_pop_plus5}
\title{Age-Stratified IpaB Model: Children >= 5 Years}
\format{
An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1200 rows and 7 columns.
}
\source{
icddr,b / UCD-SERG SOSAR cohort, Dhaka 2021-2022.
Mock version: 5 subjects, 20 iterations, 2 chains.
}
\usage{
overall_IpaB_pop_plus5
}
\description{
Age-Stratified IpaB Model: Children >= 5 Years
}
\keyword{datasets}
