% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_sensitivity_filename.R
\name{parse_sensitivity_filename}
\alias{parse_sensitivity_filename}
\title{Parse a \code{sensitivity_*.rda} file name into its components}
\usage{
parse_sensitivity_filename(path)
}
\arguments{
\item{path}{File path.}
}
\value{
One-row tibble: \verb{file, antigen, iso, prior_key, model}.
}
\description{
Expects \verb{sensitivity_<Antigen>_<Iso>_<prior>_<model>.rda}.
}
