% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_jags_output.R
\name{postprocess_jags_output}
\alias{postprocess_jags_output}
\title{Postprocess JAGS output}
\usage{
postprocess_jags_output(jags_output)
}
\arguments{
\item{jags_output}{output from \code{\link[runjags:run.jags]{runjags::run.jags()}}}
}
\value{
a \link[tibble:tbl_df-class]{tibble::tbl_df}
}
\description{
Postprocess JAGS output
}
