% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_prior_label.R
\name{recode_prior_label}
\alias{recode_prior_label}
\title{Recode the prior key to a display label}
\usage{
recode_prior_label(x)
}
\arguments{
\item{x}{Character vector of prior keys.}
}
\value{
Character vector of labels.
}
\description{
Recode the prior key to a display label
}
