% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serotype_palette.R
\name{serotype_palette}
\alias{serotype_palette}
\alias{serotype_line_alpha}
\alias{serotype_linewidth}
\title{Infecting-serotype colour palette (manuscript v2)}
\usage{
serotype_palette()

serotype_line_alpha()

serotype_linewidth()
}
\description{
Infecting-serotype colour palette (manuscript v2)
}
