% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{serotype_sf2a_3}
\alias{serotype_sf2a_3}
\alias{serotype_sf3a_3}
\alias{serotype_sonnei_3}
\title{Serotype-Specific Model Fit: S. flexneri 2a}
\format{
An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1200 rows and 7 columns.

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 400 rows and 7 columns.

An object of class \code{sr_model} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 400 rows and 7 columns.
}
\source{
icddr,b / UCD-SERG SOSAR cohort, Dhaka 2021-2022.
Mock version: 5 subjects, 20 iterations, 2 chains.
}
\usage{
serotype_sf2a_3

serotype_sf3a_3

serotype_sonnei_3
}
\description{
Hierarchical model restricted to participants infected with
\emph{S. flexneri} 2a.
}
\keyword{datasets}
