% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shigella-package.R
\docType{package}
\name{shigella-package}
\alias{shigella}
\alias{shigella-package}
\title{shigella: Shigella Antibody-Kinetics Analysis and Reporting Tools}
\description{
Reproducible analysis tools for fitting and evaluating Bayesian two-phase decay models of Shigella antibody kinetics from longitudinal serosurvey data. Provides functions for posterior prediction, model comparison, sensitivity analysis, and publication-ready tables and figures supporting serosurveillance applications.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ucd-serg.github.io/shigella/}
}

}
\author{
\strong{Maintainer}: Kwan Ho Lee \email{ksjlee@ucdavis.edu}

Authors:
\itemize{
  \item Kwan Ho Lee \email{ksjlee@ucdavis.edu}
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})
}

}
\keyword{internal}
