% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_seroincidence2.R
\name{simulate_seroincidence2}
\alias{simulate_seroincidence2}
\title{Description of the function here.}
\usage{
simulate_seroincidence2(nrep, n_sim, observed, range = NULL)
}
\arguments{
\item{nrep}{Number of repetitions.}

\item{n_sim}{Number of simulations.}

\item{observed}{Observed incidence rate.}

\item{range}{Range for simulation.}
}
\value{
A list of simulated seroincidence results.
}
\description{
Description of the function here.
}
