% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess_stan_output.R
\name{summarize_matrix_draws}
\alias{summarize_matrix_draws}
\title{Helper: summarize a draws_array of a matrix variable to a single matrix
(median across all draws)}
\usage{
summarize_matrix_draws(draws_arr, var_name, nrow, ncol)
}
\description{
Helper: summarize a draws_array of a matrix variable to a single matrix
(median across all draws)
}
