% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_study_population.R
\name{table1_study_population}
\alias{table1_study_population}
\title{Table 1: cohort characteristics (gtsummary -> flextable)}
\usage{
table1_study_population(tab1_dat)
}
\arguments{
\item{tab1_dat}{Output of \code{\link[=build_table1_data]{build_table1_data()}}.}
}
\value{
A flextable.
}
\description{
Table 1: cohort characteristics (gtsummary -> flextable)
}
