% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_s3_sensitivity.R
\name{table_s3_sensitivity}
\alias{table_s3_sensitivity}
\title{S3 Table: prior-specification robustness (overall vs serotype MAE)}
\usage{
table_s3_sensitivity(sensitivity_results)
}
\arguments{
\item{sensitivity_results}{Output of \code{\link[=build_sensitivity_results]{build_sensitivity_results()}}.}
}
\value{
A \code{gt_tbl}.
}
\description{
Data shaping is in \code{.sensitivity_wide()}; this adds gt styling.
}
