% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{load_fit_obj}
\alias{load_fit_obj}
\title{Load a Fitted Model Object from an RDA File}
\usage{
load_fit_obj(path)
}
\arguments{
\item{path}{Character scalar. Full path to the RDA file.}
}
\value{
The R object stored as \code{fit_obj} in the RDA file.
}
\description{
Load a Fitted Model Object from an RDA File
}
\examples{
\dontrun{
model <- load_fit_obj("path/to/sensitivity_Sonnei_IgG_primary_overall.rda")
}

}
