% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{recode_prior_label}
\alias{recode_prior_label}
\title{Recode Sensitivity Prior Keys to Display Labels}
\usage{
recode_prior_label(x)
}
\arguments{
\item{x}{Character vector of prior keys.}
}
\value{
A character vector with recoded values.
}
\description{
Recode Sensitivity Prior Keys to Display Labels
}
\examples{
recode_prior_label(c("primary", "diffuse", "informative", "unknown"))

}
