% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shigella-package.R
\docType{package}
\name{shigella-package}
\alias{shigella}
\alias{shigella-package}
\title{shigella: Bayesian Modeling of Shigella Antibody Kinetics}
\description{
Tools for multivariate Bayesian hierarchical modeling of antibody response trajectories following confirmed Shigella infection, supporting kinetic parameter estimation and serosurveillance applications via Stan (cmdstanr) backends.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ucd-serg.github.io/shigella/}
}

}
\author{
\strong{Maintainer}: Kwan Ho Lee \email{ksjlee@ucdavis.edu}

Authors:
\itemize{
  \item Kwan Ho Lee \email{ksjlee@ucdavis.edu}
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})
}

}
\keyword{internal}
