% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shigella-package.R
\docType{package}
\name{shigella-package}
\alias{shigella}
\alias{shigella-package}
\title{shigella: Modeling Shigella Antibody Kinetics Using Hierarchical Bayesian Methods}
\description{
Tools for modeling longitudinal antibody kinetics following Shigella infection, developed to support Chapter 1 of a doctoral dissertation. The package includes functions to run hierarchical Bayesian models using JAGS, extract posterior summaries, and visualize decay curves across multiple antigens and isotypes. It extends and reuses components from the 'serodynamics' and 'serocalculator' packages, tailored specifically to the analysis needs of Shigella seroepidemiology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ucd-serg.github.io/shigella/}
}

}
\author{
\strong{Maintainer}: Kwan Ho Lee \email{ksjlee@ucdavis.edu}

Authors:
\itemize{
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})
}

}
\keyword{internal}
