% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shigella-package.R
\docType{package}
\name{shigella-package}
\alias{shigella}
\alias{shigella-package}
\title{shigella: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kwan Ho Lee \email{ksjlee@ucdavis.edu}

Authors:
\itemize{
  \item Douglas Ezra Morrison \email{demorrison@ucdavis.edu} (\href{https://orcid.org/0000-0002-7195-830X}{ORCID})
}

}
\keyword{internal}
